update videos update

+++ modules/core_videos/module.videos.php
@@ -100,11 +100,15 @@ function videos_get_thumbnail($url) {
 		$yttlocal="$RFS_SITE_PATH/modules/core_videos/cache/$ytthumb.jpg";
 		$ytturl="$RFS_SITE_URL/modules/core_videos/cache/$ytthumb.jpg";
 		if(!file_exists($yttlocal)) {
+			if(function_exists("curl_init")) {
 			$ch = curl_init("http://i1.ytimg.com/vi/$ytthumb/mqdefault.jpg");
 			curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
 			$ytt = curl_exec($ch);
 			curl_close($ch);
 			file_put_contents("$yttlocal", $ytt);
+			
+			}
+			
 		}
 		if(!file_exists($yttlocal)) {
 			$ytturl="$RFS_SITE_URL/modules/core_videos/cache/oops.png";
